+++ Meraki.Api/Data/OrganizationWirelessMqttSettingsNetwork.cs
@@ -1,0 +1,9 @@
+namespace Meraki.Api.Data;
+
+/// <summary>
+/// Network information for wireless MQTT settings
+/// </summary>
+[DataContract]
+public class OrganizationWirelessMqttSettingsNetwork : NamedIdentifiedItem
+{
+}