+++ Meraki.Api/Data/OrganizationWirelessMqttSettingsWifiHysteresis.cs
@@ -1,0 +1,22 @@
+namespace Meraki.Api.Data;
+
+/// <summary>
+/// Wi-Fi hysteresis settings
+/// </summary>
+[DataContract]
+public class OrganizationWirelessMqttSettingsWifiHysteresis
+{
+	/// <summary>
+	/// Wi-Fi Hysteresis Enabled
+	/// </summary>
+	[ApiAccess(ApiAccess.ReadUpdate)]
+	[DataMember(Name = "enabled")]
+	public bool? Enabled { get; set; }
+
+	/// <summary>
+	/// Wi-Fi Threshold. Will be between 1 and 2147483647. Default is 1 second
+	/// </summary>
+	[ApiAccess(ApiAccess.ReadUpdate)]
+	[DataMember(Name = "threshold")]
+	public int? Threshold { get; set; }
+}