+++ Meraki.Api/Sections/General/Organizations/OrganizationsWirelessSection.cs
@@ -13,5 +13,7 @@ public class OrganizationsWirelessSection
 
 	public IOrganizationsWirelessOpportunisticPcap OpportunisticPcap { get; internal set; } = null!;
 
+	public IOrganizationsWirelessMqtt Mqtt { get; internal set; } = null!;
+
 	public OrganizationsWirelessSsidsSection Ssids { get; internal set; } = new();
 }